+++ src/qianfan/resources/auth/oauth.py
