+++ src/qianfan/tests/auth_test.py
@@ -21,7 +21,7 @@ import os
 import pytest
 
 import qianfan
-from qianfan.resources.auth import Auth
+from qianfan.resources.auth.oauth import Auth
 from qianfan.tests.utils import EnvHelper, init_test_env
 from qianfan.tests.utils.mock_server import fake_access_token
